backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
